RRDTool DB Creation and Graphing Files
The "Create_RRDTool..." file is the file used to create the database the python script will write too.
To graph the data inserted into the RRD databases, use the scripts provided. all.sh will call all scripts to graph in one shot.
StationList.txt is an example of how stations are called by the program.